Organized and updated documentation
